testing different statements

+++ srcs/vm/funcs/0x13_lld.c
@@ -6,7 +6,7 @@
 /*   By: seronen <seronen@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/06 15:09:22 by seronen           #+#    #+#             */
-/*   Updated: 2021/02/16 02:35:56 by seronen          ###   ########.fr       */
+/*   Updated: 2021/02/16 14:30:58 by seronen          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,8 +22,7 @@ void    op_lld(t_carriage *carriage, unsigned char *arena)
 		carriage->regs[reg_slot] = carriage->stmt->args[0];
 	else if (carriage->stmt->arg_types[0] == T_IND)
 	{
-		// arg_idx_value = (short)carriage->stmt->args[0] % MEM_SIZE;
-        arg_idx_value = real_modulo(carriage->abs_pos, (short)carriage->stmt->args[0], MEM_SIZE);
+        arg_idx_value = real_modulo(carriage->abs_pos, carriage->stmt->args[0], MEM_SIZE);
 		carriage->regs[reg_slot] = read_bytes(arena, arg_idx_value, 4);
 	}
 	if (carriage->regs[reg_slot] == 0)
